Realmscraft 2.0
Not a lot of technical differences in this one, but there has been a major reorganization of the functions. Lots of old stuff cut out, lots of (((subfolders))). Combine that with a consistent snake naming scheme from minecraft, and hopefully this should make it a little easier to navigate this mess.
Additionally, I went back and added some comments and basic documentation, but I'm not really sure how useful it is. 

There were a few "fixes"
*Repair armor is clears extra armor pieces, and won't override items you're wearing from outside your class >:(
*Most class items are harder to drop
*The picking up and moving of corpses has had a minor overhaul. It should be harder to stack corpses, and easier to "unstack" them
*handful of smaller fixes